requirement.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==4.2.15
django-braces==1.15.0
django-crispy-forms==2.3
sqlparse==0.5.1
typing_extensions==4.12.2
django-debug-toolbar==4.3.0